refactor: eliminate unnecessary DKG code
