--- fedimintd/src/bin/distributedgen.rs
@@ -1,6 +1,0 @@
-use fedimintd::distributed_gen::DistributedGen;
-
-#[tokio::main]
-async fn main() -> anyhow::Result<()> {
-    DistributedGen::new()?.with_default_modules().run().await
-}
